requirements.txt
mysql-connector-python>=8.0
tk>=0.1
Pillow>=9.0
